# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Chores Your Own Adventure API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: registerAdult
      tags: [auth]
      summary: Register adult account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterInput"
      responses:
        "201":
          description: Registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResult"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Email already in use
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/login:
    post:
      operationId: loginAdult
      tags: [auth]
      summary: Login adult account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResult"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/forgot-password:
    post:
      operationId: forgotPassword
      tags: [auth]
      summary: Send password reset email
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ForgotPasswordInput"
      responses:
        "200":
          description: Reset email sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /auth/reset-password:
    post:
      operationId: resetPassword
      tags: [auth]
      summary: Reset password with token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResetPasswordInput"
      responses:
        "200":
          description: Password reset
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "400":
          description: Invalid or expired token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/change-password:
    post:
      operationId: changePassword
      tags: [auth]
      summary: Change current user password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChangePasswordInput"
      responses:
        "200":
          description: Password changed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "400":
          description: Invalid current password
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /auth/kid-login:
    post:
      operationId: loginKid
      tags: [auth]
      summary: Login kid with household code and PIN
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/KidLoginInput"
      responses:
        "200":
          description: Kid logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResult"
        "401":
          description: Invalid PIN or locked
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "429":
          description: PIN attempts locked
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PinLockResponse"
  /auth/change-pin:
    post:
      operationId: changePIN
      tags: [auth]
      summary: Change kid PIN
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChangePINInput"
      responses:
        "200":
          description: PIN changed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "400":
          description: Invalid old PIN
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ─── HOME ────────────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
    MessageResponse:
      type: object
      required: [message]
      properties:
        message:
          type: string

    # ─── AUTH SCHEMAS ──────────────────────────────────────────────────────────
    RegisterInput:
      type: object
      required: [email, password, displayName]
      properties:
        email:
          type: string
        password:
          type: string
          minLength: 8
        displayName:
          type: string
          minLength: 1
    LoginInput:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
        password:
          type: string
    KidLoginInput:
      type: object
      required: [householdCode, userId, pin]
      properties:
        householdCode:
          type: string
        userId:
          type: number
        pin:
          type: string
          minLength: 4
          maxLength: 4
    ForgotPasswordInput:
      type: object
      required: [email]
      properties:
        email:
          type: string
    ResetPasswordInput:
      type: object
      required: [token, password]
      properties:
        token:
          type: string
        password:
          type: string
          minLength: 8
    ChangePasswordInput:
      type: object
      required: [currentPassword, newPassword]
      properties:
        currentPassword:
          type: string
        newPassword:
          type: string
          minLength: 8
    ChangePINInput:
      type: object
      required: [oldPin, newPin]
      properties:
        oldPin:
          type: string
          minLength: 4
          maxLength: 4
        newPin:
          type: string
          minLength: 4
          maxLength: 4
    AuthResult:
      type: object
      required: [user, token]
      properties:
        user:
          $ref: "#/components/schemas/UserProfile"
        token:
          type: string
    UserProfile:
      type: object
      required: [id, displayName, userType, isAppAdmin, lifetimeXp, currentLevel, personalGold, adventureMode, soundEnabled, hapticsEnabled]
      properties:
        id:
          type: number
        email:
          type: ["string", "null"]
        displayName:
          type: string
        adventurerName:
          type: ["string", "null"]
        userType:
          type: string
          enum: [adult, kid]
        isAppAdmin:
          type: boolean
        lifetimeXp:
          type: number
        currentLevel:
          type: number
        highestLevel:
          type: number
        personalGold:
          type: number
        legendaryCompletions:
          type: number
        adventureMode:
          type: boolean
        soundEnabled:
          type: boolean
        hapticsEnabled:
          type: boolean
        characterCooldownUntil:
          type: ["string", "null"]
        activePartyId:
          type: ["number", "null"]
        createdAt:
          type: string
...
    PinLockResponse:
      type: object
      required: [error, lockedUntil]
      properties:
        error:
          type: string
        lockedUntil:
          type: string

    # ─── HOME SCHEMAS ──────────────────────────────────────────────────────────
...