# lib/api-zod/src/generated/types/authResult.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Chores Your Own Adventure API
 * OpenAPI spec version: 0.1.0
 */
import type { UserProfile } from './userProfile';

export interface AuthResult {
  user: UserProfile;
  token: string;
}